+++ spec/rubocop/cop/rails/unique_validation_without_index_spec.rb
@@ -495,5 +495,44 @@ RSpec.describe RuboCop::Cop::Rails::UniqueValidationWithoutIndex, :config do
         end
       end
     end
+
+    context 'when the table does not exist' do
+      let(:schema) { <<~RUBY }
+        ActiveRecord::Schema.define(version: 2020_02_02_075409) do
+          create_table "users", force: :cascade do |t|
+            t.string "account", null: false, unique: true
+          end
+        end
+      RUBY
+
+      it 'does not register an offense' do
+        expect_no_offenses(<<~RUBY)
+          class Article
+            validates :account, uniqueness: true
+          end
+        RUBY
+      end
+    end
+
+    context 'when module' do
+      let(:schema) { <<~RUBY }
+        ActiveRecord::Schema.define(version: 2020_02_02_075409) do
+          create_table "users", force: :cascade do |t|
+            t.string "account", null: false, unique: true
+          end
+        end
+      RUBY
+
+      it 'does not register an offense' do
+        expect_no_offenses(<<~RUBY)
+          module User
+            extend ActiveSupport::Concern
+            included do
+              validates :account, uniqueness: true
+            end
+          end
+        RUBY
+      end
+    end
   end
 end